requirements.txt
langchain
langchain-core
langchain-community
langchain-openai

openai
tiktoken

faiss-cpu
pymilvus
langchain-milvus

pypdf
python-docx
python-pptx
pandas
openpyxl
unstructured

python-dotenv
numpy

streamlit
